Turn AdvancedDetailsModal into a card - Move My Receive Address into the card - Move Transaction ID out of the card
